web/requirements/base.txt
Django
psycopg2 # для подключения к постгре
djangorestframework
django-rest-auth[with_social]==0.9.5
djangorestframework-jwt
django-summernote # текстовой редактор
django-allauth==0.42.0
pillow # для работы с изображениями
drf-yasg
